Move tests to new namespace
